+++ shared/chat/conversation/input-area/normal/index.js
@@ -3,7 +3,7 @@ import * as React from 'react'
 import {Input as TextInput} from '../../../../common-adapters'
 import * as Types from '../../../../constants/types/chat2'
 import {isMobile} from '../../../../util/container'
-import MentionInput from './mention-handler-hoc'
+import MentionInput from './mention-input'
 import {type InputProps} from './platform-index'
 import {throttle} from 'lodash-es'
 import {formatTextForQuoting} from '../../../../util/chat'
